Move `stat` utils to `fs-utils`
