+++ packages/pridepack/src/core/fs-utils.ts
@@ -8,11 +8,21 @@ export async function removeFile(filePath: string): Promise<void> {
   return fs.rm(filePath, { recursive: true, force: true });
 }
 
-export async function fileExists(filePath: string): Promise<boolean> {
+
+export async function isFile(filePath: string): Promise<boolean> {
   try {
     const stat = await fs.stat(filePath);
     return stat.isFile();
-  } catch (error) {
+  } catch (err) {
+    return false;
+  }
+}
+
+export async function isDirectory(filePath: string): Promise<boolean> {
+  try {
+    const stat = await fs.stat(filePath);
+    return stat.isDirectory();
+  } catch (err) {
     return false;
   }
 }
